Add duplicate handling (unique index: simp + raw_pinyin)
Ok: so the previous problem was that there are duplicate characters (naturally occurring in Chinese),
and the queries in the Rust+mongoDB driver only pull the first match (which makes sense). However, this
can lead to less accurate lookups for articles. Previously this problem was mitigated by merging adjacent
cases in CEDICT where one entry matched the previous, however not all cases are adjacent.

In this commit, I modified the workflow to first sort the CEDICT input and then output as a csv.
This guarantees that similar phrases are adjacent (at least by what I sorted on: simp + raw_pinyin).
This approach seems mostly ok, however it's loses the cases where a traditional char maps to multiple
simplified chars. I am prioritizing simplified for now and will revisit this later if necessary.

Also added the <hr> tag to split multiple definitions. Neat!