Fix url upload bug
Turns out the individual components worked correctly,
however the socket was inconsistently streaming the data.
My guess is this has to do with some async funny business.

The current resolution is to add sleep commands on the
Rust function after write and after the read to ensure that
everything gets written. 1s is picked as a reasonable margin,
though change as-needed.