Add initial Redis implementation
Alright this is a big one -- moved all CEDICT lookups out of mongoDB and into Redis.

In order to do this, the following changes were made:
1) Lookups are done using the Simplified Phrase + Raw Pinyin as the unique id (uid)
- These are guaranteed to be unique during the CEDICT load into Redis
- tokenserver.py was updated to only send these two during tokenization
- - An additional library (opencc) is used to convert traditional->simplified before tokenization
- - Simplified was chosen because it was slightly less lossy than Traditional with opencc conversions
2) MongoDB was moved back to Sync (from Async)
- This is cleaner code-wise for approx same performance.
- The technical reason was Tokio versions were conflicting.
3) The URL parsing call (reqwest) was made async.
- Here, the blocking version of reqwest uses Tokio under the hood which affected the outside runtime

For reference, the file "redis_cedict_keys.txt" is provided as reference which has the uids used to
store the corresponding data (using HSETs).

There are still some bugs to clean-up, though this has improved render times.

Will need to clean-up bugs before pushing to PRD