Actually fix url upload issue
The actual issue was: for large packets the socket sends it as blocks \
of data (which is standard), whereas I assumed it sent as a continuous stream.

The fix: send a fixed-length header message to communicate the size
of the payload. This is 64 bits. Then read exactly that amount into the
resulting buffer. Thus the stream will only move-on after reading
the appropriate amount of data (or there's an error).